Makse resample_ds function and update retrieve_data.
The retrieve_data function is updated. It is now able to resample datasets found in the ESGF database if the datasets have different frequency than the frequency of the cutout, defined as 'dt'.

This is done by creating a new function, resample_ds, which works for the case of resampling from frequencies of 6hr and daily to a frequency of 3hr. The resampling method used is ds.resample().nearest().

The resample_ds function is not able to resample daily values starting at other times than 12 o'clock. If such a function is wanted / needed, the resample_ds function must be updated.